+++ src/main/java/net/thedragonteam/armorplus/container/SlotIngredient.java
@@ -26,11 +26,6 @@ public class SlotIngredient extends SlotItemHandler {
     }
 
     @Override
-    public int getSlotStackLimit() {
-        return 1;
-    }
-
-    @Override
     public int getItemStackLimit(@Nonnull ItemStack stack) {
         return 1;
     }