+++ src/main/java/net/thedragonteam/armorplus/container/SlotOutput.java
@@ -8,8 +8,6 @@ import net.minecraft.item.ItemStack;
 import net.minecraftforge.items.IItemHandler;
 import net.minecraftforge.items.SlotItemHandler;
 
-import javax.annotation.Nonnull;
-
 public class SlotOutput extends SlotItemHandler {
 
     public SlotOutput(IItemHandler itemHandler, int index, int xPosition, int yPosition) {
@@ -19,10 +17,5 @@ public class SlotOutput extends SlotItemHandler {
     @Override
     public boolean isItemValid(ItemStack stack) {
         return false;
-    }
-
-    @Override
-    public int getItemStackLimit(@Nonnull ItemStack stack) {
-        return 1;
     }
 }